* multiple bugfixes with calculation dashboard; * added a button to open results file; * added a button to open photo folder.
Former-commit-id: 16274882a11f6c44a3daf2e0ef01a532eccb0bab
Former-commit-id: ff4ef19d77f70883adf7a8304fbd02f793c5eb62

+++ Primitives/CalculationResult.cs
@@ -18,12 +18,12 @@ namespace Primitives
 
 		public int ObjectHeight { get; }
 
-		public long ObjectVolume { get; }
+		public double ObjectVolume { get; }
 
 		public string CalculationComment { get; }
 
 		public CalculationResult(DateTime calculationTime, string objectCode, double objectWeight, uint unitCount,
-			int objectLength, int objectWidth, int objectHeight, long objectVolume, string calculationComment)
+			int objectLength, int objectWidth, int objectHeight, double objectVolume, string calculationComment)
 		{
 			CalculationTime = calculationTime;
 			ObjectCode = objectCode;
